fix: unification rules
