+++ src/index.ts
@@ -1,6 +1,5 @@
 module.exports = {
   rules: {
     "no-excess-property": require("./no-excess-property"),
-    "no-excess-property-func": require("./no-excess-property-func"),
   },
 };
