+++ src/no-excess-property.ts
@@ -1,6 +1,6 @@
-import { ESLintUtils } from "@typescript-eslint/utils";
 import ts from "typescript";
-import { TypeUtil } from "./util";
+import { ESLintUtils, TSESTree } from "@typescript-eslint/utils";
+import { Checker, TypeUtil } from "./util";
 
 const createRule = ESLintUtils.RuleCreator(
   (name) => `https://npmjs.com/package/${name}`
@@ -9,15 +9,48 @@ const createRule = ESLintUtils.RuleCreator(
 export = createRule({
   create(context) {
     const parserServices = context.sourceCode.parserServices;
-    const checker =
-      parserServices?.program?.getTypeChecker() as ts.TypeChecker & {
-        getElementTypeOfArrayType: (type: ts.Type) => ts.Type | undefined;
-      };
+    const checker = parserServices?.program?.getTypeChecker() as Checker;
 
     if (!checker || !parserServices) return {};
-    const typeUtil = new TypeUtil(checker);
+    const typeUtil = new TypeUtil(checker, parserServices);
+
+    const checkReturnStatement = (
+      returnStatements: TSESTree.ReturnStatement[],
+      returnTypes: ts.Type[]
+    ) => {
+      for (const returnStatement of returnStatements) {
+        if (!returnStatement.argument) return;
+        const returnStatementRawNode = checker.getTypeAtLocation(
+          parserServices.esTreeNodeToTSNodeMap!.get(returnStatement.argument)
+        );
+        const returnStatementNode =
+          checker.getPromisedTypeOfPromise(returnStatementRawNode) ??
+          returnStatementRawNode;
+
+        if (
+          returnTypes.every((type) =>
+            typeUtil.checkProperties(returnStatementNode, type, true)
+          )
+        ) {
+          context.report({
+            node: returnStatement,
+            messageId: "no-excess-property-func",
+          });
+        }
+      }
+    };
 
     return {
+      FunctionDeclaration(node) {
+        const returnTypes = typeUtil.getReturnTypes(node);
+        if (returnTypes.length > 0) {
+          const returnStatements = node.body.body
+            .map((e) => typeUtil.findReturnStatements(e))
+            .flat();
+          checkReturnStatement(returnStatements, returnTypes);
+        }
+      },
+
       CallExpression(node) {
         for (const [idx, argument] of node.arguments.entries()) {
           const argNode = parserServices.esTreeNodeToTSNodeMap!.get(argument);
@@ -55,6 +88,41 @@ export = createRule({
       },
 
       VariableDeclarator(node) {
+        const returnTypes = typeUtil.getReturnTypes(node);
+
+        if (returnTypes.length > 0) {
+          if (
+            node.init !== null &&
+            "body" in node.init &&
+            node.init.body.type === "BlockStatement"
+          ) {
+            const returnStatements = node.init.body.body
+              .map((e) => typeUtil.findReturnStatements(e))
+              .flat();
+            checkReturnStatement(returnStatements, returnTypes);
+            return;
+          }
+          if (node.init !== null && "body" in node.init) {
+            const returnStatementRawNode = checker.getTypeAtLocation(
+              parserServices.esTreeNodeToTSNodeMap!.get(node.init.body)
+            );
+            const returnStatementNode =
+              checker.getPromisedTypeOfPromise(returnStatementRawNode) ??
+              returnStatementRawNode;
+
+            if (
+              returnTypes.every((type) =>
+                typeUtil.checkProperties(returnStatementNode, type, true)
+              )
+            ) {
+              context.report({
+                node,
+                messageId: "no-excess-property-func",
+              });
+            }
+          }
+        }
+
         if (
           !node.init ||
           !node.id.typeAnnotation ||
@@ -96,6 +164,7 @@ export = createRule({
     messages: {
       "no-excess-property":
         "Object has property not present {{ excessProperty }} in type",
+      "no-excess-property-func": "Object has property not present in type",
     },
     schema: [],
   },
